Repositorios y conexion a la base de datos

+++ src/main/java/com/alfonso/empleos/service/ICategoriasService.java
@@ -7,5 +7,5 @@ import com.alfonso.empleos.model.Categoria;
 public interface ICategoriasService {
 	void guardar(Categoria categoria);
 	List<Categoria> buscarTodas();
-	Categoria buscarPorId(Integer id);
+	Categoria buscarPorId(Long id);
 }
